requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
fastapi==0.95.0
uvicorn==0.22.0
tensorflow==2.18.0
numpy
pandas
scikit-learn
httpx
